cli: Utilise un long plutôt qu'un int pour le code de statut HTTP
Curl s'attend à un pointeur vers un long, ce qui pouvait causer sur certains systèmes une erreur de segmentation :
Curl modifiait précédemment potentiellement des zones mémoires en dehors du type entier, ce qui pouvait altérer les variables environnantes, comme contentType.

+++ cli/src/io/HttpRequest.cpp
@@ -74,7 +74,7 @@ namespace io {
         if (res != CURLE_OK) {
             throw RequestException{static_cast<int>(res)};
         }
-        int httpStatus = 0;
+        long httpStatus = 0;
         curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &httpStatus);
         return HttpResponse{httpStatus, std::move(contentType), std::move(buffer)};
     }
@@ -103,10 +103,10 @@ namespace io {
         curl_easy_cleanup(curl);
     }
 
-    HttpResponse::HttpResponse(int statusCode, std::string contentType, std::string content)
+    HttpResponse::HttpResponse(long statusCode, std::string contentType, std::string content)
         : statusCode{statusCode}, contentType{std::move(contentType)}, content{std::move(content)} {}
 
-    int HttpResponse::getStatusCode() const {
+    long HttpResponse::getStatusCode() const {
         return statusCode;
     }
